Follow hybrid /XRefStm, resolve indirect filters, harden crypto and numbers
- ParseClassicXref follows the trailer /XRefStm to the cross-reference
  stream so compressed objects in hybrid-reference files resolve.
- Filter dispatch resolves indirect /Filter array items and indirect
  /DecodeParms (including inside object streams) before decoding.
- AesCbc.Decrypt tolerates ciphertext that is not a whole number of
  blocks by dropping the trailing partial block.
- Revision 6 passwords are encoded as UTF-8; /StmF and /StrF select the
  crypt filter by name from /CF, honoring /Identity and named filters.
- The repair path preserves /Encrypt, /ID and /Info from the last raw
  trailer, and the object cache is cleared once decryption is ready.
- NumberObject keeps 64-bit integers intact and refuses to serialize
  NaN or infinite reals.

+++ Radzen.Blazor/Documents/Pdf/Objects/ObjectParser.cs
@@ -66,7 +66,7 @@ internal sealed class ObjectParser
             return new ReferenceObject((int)token.IntValue, (int)generation.IntValue);
         }
 
-        return new NumberObject((int)token.IntValue);
+        return new NumberObject(token.IntValue);
     }
 
     private DocumentObject ParseArray()
